[refactor] organize seperated test files into Default/Actions/Scripts test files
