+++ TestConsole/src/test/java/ActionsTest.java
@@ -1,14 +1,47 @@
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.List;
 
+import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.anyInt;
 import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.times;
 
-class WriteTest {
+public class ActionsTest {
+    @Test
+    void 읽기_1_LBA_성공() {
+        int input = 3;
+        String expect = "0xAAAABBBB";
+
+        TestConsole mockConsole = mock(TestConsole.class);
+        when(mockConsole.read(input)).thenReturn(expect);
+
+        assertThat(mockConsole.read(input)).isEqualTo(expect);
+    }
+
+    @Test
+    void 읽기_1_LBA_실패() {
+        int input = 4;
+        String expect = "ERROR";
+
+        TestConsole mockConsole = mock(TestConsole.class);
+        when(mockConsole.read(input)).thenReturn(expect);
+
+        assertThat(mockConsole.read(input)).startsWith(expect);
+    }
+
+    @Test
+    void 읽기_FULL_LBA_성공() {
+        TestConsole mockConsole = mock(TestConsole.class);
+
+        mockConsole.fullRead();
+
+        verify(mockConsole, times(100)).read(anyInt());
+    }
 
     @Test
     void WRITE에서_익셉션안나면_성공() {
